Refactor and add new frames in Basketball Tracker UI
The Basketball Tracker User Interface was significantly restructured. New frames, including ControlFrame, InputFrame, and OutputFrame, were added into the MainWindow to improve the organization of the application. The MainWindow is now fixed in the center of the screen. The StateManager class's constructor now takes a signal distributor instance, extending its functionality. This set of changes improves the usability and readability of the Basketball Tracker application.