Refactor database queries and fix minor issues
The database queries have been refactored to use context managers for better handling of resource cleanup. The table names have also been updated in these queries to better represent the source of the information. Some minor issues such as a typo in the ControlFrame class and an unnecessary file in the .gitignore file have also been addressed.