Refactor data access in OutputFrame class
Three Data Access Objects (DAOs) - RostersDAO, EventsDAO, and PlayerStatsDAO have been refactored in the OutputFrame class. The classes are initialized in the constructor instead of being initialized in each method. This should improve the performance of the database loads in the tabs since the DAOs are now shared across methods instead of being re-created each time.